Adding security part - 1

+++ backend_microservice/review-eventmicroservices/src/main/java/com/locality/review/eventmicroservices/service/impl/ReviewServiceImpl.java
@@ -5,13 +5,14 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.locality.review.eventmicroservices.entity.Review;
+import com.locality.review.eventmicroservices.exception.NotAuthorizedException;
 import com.locality.review.eventmicroservices.exception.ResourceExistsException;
 import com.locality.review.eventmicroservices.exception.ResourceNotFoundException;
+import com.locality.review.eventmicroservices.mapper.ReviewMapper;
 import com.locality.review.eventmicroservices.payload.LocalityAndEventTypeDto;
 import com.locality.review.eventmicroservices.payload.ReviewDto;
 import com.locality.review.eventmicroservices.payload.UserDto;
@@ -34,22 +35,21 @@ public class ReviewServiceImpl implements ReviewService {
 	private FetchService fetchService;
 	
 	@Autowired
-	private ModelMapper modelMapper;
+	private ReviewMapper reviewMapper;
 
 	@Override
-	public ReviewDto saveReview(ReviewDto reviewDto, Long userId, Long localityId) 
+	public ReviewDto saveReview(ReviewDto reviewDto, String token, Long localityId) 
 			throws ResourceNotFoundException, ResourceExistsException{
 		
-		UserDto user = this.fetchService.getUser(userId);
+		UserDto user = this.fetchService.validateUser(token);
 		LocalityAndEventTypeDto locality = this.fetchService.getLocality(localityId);
 		
 
-		if (!this.reviewRepository.findByUserIdAndLocalityIdOrderByDate(userId, localityId).isEmpty()) {
-			log.warn("Duplicate review");
-			throw new ResourceExistsException("You have already added a review for the selected locality");
+		if (!this.reviewRepository.findByUserIdAndLocalityIdOrderByDate(user.getUserId(), localityId).isEmpty()) {
+			throw new ResourceExistsException("Duplicate review");
 		}
 		
-		Review review = this.dtoToReview(reviewDto);
+		Review review = reviewMapper.dtoToReview(reviewDto);
 
 		log.info("Saving new review");
 		review.setImg((int) (Math.random() * 4) + 1);
@@ -58,7 +58,7 @@ public class ReviewServiceImpl implements ReviewService {
 		review.setUsername(user.getUsername());
 		review.setLocalityId(locality.getLocalityId());
 		review.setLocalityname(locality.getName());
-		return this.reviewToDto(this.reviewRepository.save(review));
+		return reviewMapper.reviewToDto(this.reviewRepository.save(review));
 	}
 
 	@Override
@@ -67,12 +67,11 @@ public class ReviewServiceImpl implements ReviewService {
 		List<Review> reviews = this.reviewRepository.findByOrderByDate();
 
 		if (reviews.isEmpty()) {
-			log.error("Reviews not found");
 			throw new ResourceNotFoundException("Reviews not found");
 		}
 		
 		List<ReviewDto> reviewDtos = reviews.stream()
-				.map(review -> this.reviewToDto(review)).collect(Collectors.toList());
+				.map(review -> reviewMapper.reviewToDto(review)).collect(Collectors.toList());
 		
 		return reviewDtos;
 	}
@@ -83,12 +82,11 @@ public class ReviewServiceImpl implements ReviewService {
 		List<Review> reviews = this.reviewRepository.findTop10ByOrderByDate();
 
 		if (reviews.isEmpty()) {
-			log.error("Reviews not found");
 			throw new ResourceNotFoundException("Reviews not found");
 		}
 
 		List<ReviewDto> reviewDtos = reviews.stream()
-				.map(review -> this.reviewToDto(review)).collect(Collectors.toList());
+				.map(review -> reviewMapper.reviewToDto(review)).collect(Collectors.toList());
 		
 		return reviewDtos;
 	}
@@ -97,92 +95,105 @@ public class ReviewServiceImpl implements ReviewService {
 	public List<ReviewDto> getAllReviewByLocality(Long localityId) throws ResourceNotFoundException,
 			IllegalArgumentException{
 
-		this.fetchService.getLocality(localityId);
-
 		List<Review> reviews = reviewRepository.findByLocalityIdOrderByDate(localityId);
 
 		if (reviews.isEmpty()) {
-			log.error("Reviews not found for locality id "+localityId);
-			throw new ResourceNotFoundException("Reviews not found for locality id "+localityId);
+			throw new ResourceNotFoundException("Reviews not found for selected locality");
 		}
 		
 		List<ReviewDto> reviewDtos = reviews.stream()
-				.map(review -> this.reviewToDto(review)).collect(Collectors.toList());
+				.map(review -> reviewMapper.reviewToDto(review)).collect(Collectors.toList());
 		
 		return reviewDtos;
 	}
 
 	@Override
-	public List<ReviewDto> getAllReviewByUser(Long userId) throws ResourceNotFoundException,
+	public List<ReviewDto> getAllReviewByUser(String token) throws ResourceNotFoundException,
 		IllegalArgumentException{
-
-		//System.out.println(this.fetchService.getUser(userId));
-
-		List<Review> reviews = reviewRepository.findByUserIdOrderByDate(userId);
+		
+		UserDto user = this.fetchService.validateUser(token);
+		List<Review> reviews = reviewRepository.findByUserIdOrderByDate(user.getUserId());
 
 		if (reviews.isEmpty()) {
-			log.error("Reviews not found for user with id "+userId);
-			throw new ResourceNotFoundException("Reviews not found for user with id "+userId);
+			throw new ResourceNotFoundException("Reviews not found");
 		}
 		
 		List<ReviewDto> reviewDtos = reviews.stream()
-				.map(review -> this.reviewToDto(review)).collect(Collectors.toList());
+				.map(review -> reviewMapper.reviewToDto(review)).collect(Collectors.toList());
+		
+		return reviewDtos;
+	}
+	
+	
+	@Override
+	public List<ReviewDto> getAllReviewByUserAndLocality(String token, Long localityId) 
+			throws ResourceNotFoundException{
+		
+		UserDto user = this.fetchService.validateUser(token);
+		List<Review> reviews = this.reviewRepository
+				.findByUserIdAndLocalityIdOrderByDate(user.getUserId(), localityId);
+		
+		if (reviews.isEmpty()) {
+			throw new ResourceNotFoundException("Reviews not found");
+		}
+		
+		List<ReviewDto> reviewDtos = reviews.stream()
+				.map(review -> reviewMapper.reviewToDto(review)).collect(Collectors.toList());
 		
 		return reviewDtos;
 	}
 
+	
 	@Override
-	public ReviewDto updateReview(ReviewDto reviewDto, Long reviewId) {
+	public ReviewDto updateReview(ReviewDto reviewDto, Long reviewId, String token) 
+	throws ResourceNotFoundException, IllegalArgumentException, NotAuthorizedException{
+		
+		UserDto validateUser = this.fetchService.validateUser(token);
 		
 		if(reviewDto == null) {
 			throw new IllegalArgumentException("No review body found for update");
 		}
 
 		Optional<Review> searchedReview = reviewRepository.findById(reviewId);
-
+		
+		
 		if (searchedReview.isEmpty()) {
-			log.error("Review not found for review id "+reviewId);
-			throw new ResourceNotFoundException("Review not found for review id "+reviewId);
+			throw new ResourceNotFoundException("Review not found");
 		}
+		
+		if(searchedReview.get().getUserId() != validateUser.getUserId()) {
+			throw new NotAuthorizedException("User is unauthorized");
+		}
+
 
 		Review updatedReview = searchedReview.get();
 		
 		updatedReview.setContent(reviewDto.getContent() == null ?
 				updatedReview.getContent() : reviewDto.getContent());
 
-		return this.reviewToDto(this.reviewRepository.save(updatedReview));
+		return reviewMapper.reviewToDto(this.reviewRepository.save(updatedReview));
 	}
 
 	@Override
-	public boolean deleteReview(Long reviewId) throws ResourceNotFoundException{
+	public Boolean deleteReview(Long reviewId, String token) throws ResourceNotFoundException,
+	NotAuthorizedException{
+		
+		UserDto validateUser = this.fetchService.validateUser(token);
+		
 		Optional<Review> doesReviewExist = reviewRepository.findById(reviewId);
 
 		if (doesReviewExist.isEmpty()) {
-			log.error("Review not found for review id "+reviewId);
-			throw new ResourceNotFoundException("Review not found for review id "+reviewId);
+			throw new ResourceNotFoundException("Review not found");
 		}
-
+		
+		if(doesReviewExist.get().getUserId() != validateUser.getUserId()) {
+			throw new NotAuthorizedException("User is unauthorized");
+		}
+		
 		reviewRepository.deleteById(reviewId);
 		return true;
 	}
 
 
-	@Override
-	public Review dtoToReview(ReviewDto reviewDto) {
-		return this.modelMapper.map(reviewDto, Review.class);
-	}
 
-	@Override
-	public ReviewDto reviewToDto(Review review) {
-		
-		return ReviewDto.buildReviewDto(review.getReviewId(), 
-				review.getDate(),
-				review.getImg(),
-				review.getContent(),
-				null,
-				review.getUsername(),
-				null,
-				review.getLocalityname());
-		
-	}
 }
